+++ src/main/java/insane96mcp/enhancedai/modules/Modules.java
@@ -9,6 +9,7 @@ import insane96mcp.enhancedai.modules.enderman.EndermanModule;
 import insane96mcp.enhancedai.modules.ghast.GhastModule;
 import insane96mcp.enhancedai.modules.skeleton.SkeletonModule;
 import insane96mcp.enhancedai.modules.spider.SpiderModule;
+import insane96mcp.enhancedai.modules.villager.VillagerModule;
 import insane96mcp.enhancedai.modules.witch.WitchModule;
 import insane96mcp.enhancedai.modules.zombie.ZombieModule;
 
@@ -23,6 +24,7 @@ public class Modules {
 	public static EndermanModule enderman;
 	public static SkeletonModule skeleton;
 	public static SpiderModule spider;
+	public static VillagerModule villager;
 	public static WitchModule witch;
 	public static ZombieModule zombie;
 
@@ -36,6 +38,7 @@ public class Modules {
 		enderman = new EndermanModule();
 		skeleton = new SkeletonModule();
 		spider = new SpiderModule();
+		villager = new VillagerModule();
 		witch = new WitchModule();
 		zombie = new ZombieModule();
 	}
@@ -50,6 +53,7 @@ public class Modules {
 		enderman.loadConfig();
 		skeleton.loadConfig();
 		spider.loadConfig();
+		villager.loadConfig();
 		witch.loadConfig();
 		zombie.loadConfig();
 	}