+++ src/main/java/insane96mcp/enhancedai/modules/animal/feature/AnimalAttacking.java
@@ -92,7 +92,7 @@ public class AnimalAttacking extends Feature {
 
         CompoundTag persistentData = animal.getPersistentData();
 
-        double movementSpeedMultiplier = NBTUtils.getDoubleOrPutDefault(persistentData, EAStrings.Tags.Animal.SPEED_MULTIPLIER_WHEN_AGGROED, this.speedMultiplier);
+        double movementSpeedMultiplier = NBTUtils.getDoubleOrPutDefault(persistentData, EAStrings.Tags.Passive.SPEED_MULTIPLIER_WHEN_AGGROED, this.speedMultiplier);
 
         if (this.animalsFightBack) {
             animal.targetSelector.addGoal(1, (new HurtByTargetGoal(animal)).setAlertOthers());